% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genophenogram.R
\name{axisHack}
\alias{axisHack}
\title{Helper function to draw staggered oversized labels on the y-axis}
\usage{
axisHack(at, labels, cex = 1)
}
\arguments{
\item{at}{positions of tick marks}

\item{labels}{tick mark labels}

\item{cex}{character expansion factor (relative font size)}
}
\description{
Helper function to draw staggered oversized labels on the y-axis
}
