% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologs.R
\name{calc.conservation}
\alias{calc.conservation}
\title{Calculate position-wise conservation from a Uniprot accession}
\usage{
calc.conservation(acc, overrideCache = FALSE)
}
\arguments{
\item{acc}{the Uniprot accession}
}
\value{
a numerical vector with the position-wise conservation.
}
\description{
Retrieves the 90% most similar proteins from Uniprot, runs ClustalO to 
calculate a multiple sequence alignment and uses the AMAS algorithm to 
derive the position-wise sequence conservation. If the given accession
has been used as an input before, a cached sequence alignment will be
used instead.
}
