% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{dashboard.async.run}
\alias{dashboard.async.run}
\title{Launch asynchronous dashboard job}
\usage{
dashboard.async.run(ssid, uniprotId, pdbs, mainChains, wt.seq = NULL,
  seq.offset = 0, syn.med = NULL, stop.med = NULL,
  overrideCache = FALSE, outFormats = c("pdf", "png"), pngRes = 100)
}
\arguments{
\item{ssid}{The MaveDB ScoreSet ID of the dataset to be visualized}

\item{uniprotId}{The UniprotKB accession for the corresponding protein}

\item{pdbs}{A vector of PDB accessions to be used for the structure tracks}

\item{mainChains}{the chain identifiers corresponding to the protein in question
for each provided pdb accession.}

\item{wt.seq}{An (optional) wild-type sequence for the given protein to use instead of the
sequence found in UniprotKB. This can be either a nucleotide or amino acid sequence.}

\item{seq.offset}{A parameter describing the start position of the map in the protein's
amino acid sequence. For example, if only a domain was scanned.}

\item{syn.med}{The median value of synonymous variants as an estimate of wild-type
function. This is only necessary of no synonymous variants are present in the dataset.}

\item{stop.med}{The median value of nonsense variants as an estimate of complete loss
of function. This is only necessary if no nonsense variants are present in the dataset.}

\item{overrideCache}{defaults to \code{FALSE}. If set to \code{TRUE}, data will always
be re-downloaded from remote locations instead of using a local cache.}

\item{outFormats}{a vector containing any of the following strings: x11, pdf, png .
Using two or all three at once is allowed and will result in multiple output files.}

\item{pngRes}{Resolution of PNG output in DPI. Defaults to 100.}
}
\value{
the job ID
}
\description{
Launches a job running the dashboard function in a separate system process
in the background. The job is given a unique ID, which is returned by this function.
}
