% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavevis.R
\name{dashboard}
\alias{dashboard}
\title{Draw dashboard for scoreset}
\usage{
dashboard(ssid, uniprotId, pdbs, mainChains, wt.seq = NULL,
  seq.offset = 0, syn.med = NULL, stop.med = NULL,
  overrideCache = FALSE, outFormats = c("pdf", "png"), pngRes = 100,
  outID = ssid)
}
\arguments{
\item{ssid}{The MaveDB ScoreSet ID of the dataset to be visualized}

\item{uniprotId}{The UniprotKB accession for the corresponding protein}

\item{pdbs}{A vector of PDB accessions to be used for the structure tracks}

\item{mainChains}{the chain identifiers corresponding to the protein in question
for each provided pdb accession.}

\item{wt.seq}{An (optional) wild-type sequence for the given protein to use instead of the
sequence found in UniprotKB. This can be either a nucleotide or amino acid sequence.}

\item{seq.offset}{A parameter describing the start position of the map in the protein's
amino acid sequence. For example, if only a domain was scanned.}

\item{syn.med}{The median value of synonymous variants as an estimate of wild-type
function. This is only necessary of no synonymous variants are present in the dataset.}

\item{stop.med}{The median value of nonsense variants as an estimate of complete loss
of function. This is only necessary if no nonsense variants are present in the dataset.}

\item{overrideCache}{defaults to \code{FALSE}. If set to \code{TRUE}, data will always
be re-downloaded from remote locations instead of using a local cache.}

\item{outFormats}{a vector containing any of the following strings: x11, svg, pdf, png .
Using two or all three at once is allowed and will result in multiple output files.}

\item{pngRes}{Resolution of PNG output in DPI. Defaults to 100.}

\item{outID}{a name for the output file to which the plot will be written. Defaults to ssid.}
}
\description{
Retrieves a scoreset entry from MaveDB and renders a dashboard plot consisting
of a genophenogram heatmap, interface burial, solvent accessibility, secondary
structure and conservation. The output is controlled by the \code{outFormats}
parameter and can be either on an X11 device or in PDF, SVG or PNG format (or any 
combination thereof). Structural and conservation information is obtained from
UniProtKB and PDB, thus the respective database accessions are required. Multiple
PDB files can be used at once.
}
\examples{
\dontrun{
dashboard(
 ssid="urn:mavedb:00000001-a-1", uniprotId="P63279",
 pdbs="3UIP", mainChains="A",
  syn.med=1,stop.med=0
)
}
}
