% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavevis.R
\name{getCacheFile}
\alias{getCacheFile}
\title{Find cache file location by name}
\usage{
getCacheFile(name)
}
\arguments{
\item{name}{the name of the file, e.g. "P12456_alignment.fasta"}
}
\value{
the full path to the file
}
\description{
Finds the location for a cache file. The file does not necessary need to exist yet,
as this function is meant to be used determine to a location for both storage and retrieval.
}
\details{
Depending on the execution context, the storage location may differ. The cache location can 
be controlled with the environment variable \code{$MAVEVIS_CACHE}. This will be made use of within
the mavevis server docker container. If the variable is not set, a directory ".mavecache/"
will be created in the user's home directory to be used as the storage location.
}
\examples{
file <- getCacheFile("P12345_alignment.fasta")

}
