% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologs.R
\name{getUniprotSeq}
\alias{getUniprotSeq}
\title{Retrieve Uniprot Sequence}
\usage{
getUniprotSeq(uniprot.acc)
}
\arguments{
\item{uniprot.acc}{the accession}
}
\value{
the amino acid sequence
}
\description{
Retrieves the amino acid sequence for the protein indicated by a Uniprot accession.
}
