% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{makeUUID}
\alias{makeUUID}
\title{Create universally unique ID (UUIDv4)}
\usage{
makeUUID()
}
\value{
the UUID as a character string
}
\description{
Creates a universally unique identifier compatible with the UUID v4.0 standard.
See \url{https://en.wikipedia.org/wiki/Universally_unique_identifier#Version_4_(random)}
}
