% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amasLite.R
\name{new.amasLite}
\alias{new.amasLite}
\title{Calculate AMAS conservation}
\usage{
new.amasLite()
}
\value{
an object of type amasLite
}
\description{
Uses the AMAS method (Livingstone & Barton 1993) method to calculate position-wise
conservation from a protein multiple sequence alignment. Creates an object of type
amasLite, which features a single method: \code{run(alignmentFile)}.
}
\details{
The \code{run(alignmentFile)} method takes the name of a file as its parameter. 
The file must be a FASTA formatted multiple sequence alignment. The method returns
a numerical vector listing the position-wise conservation with respect to the first
entry in the alignment.
}
