% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libpdb.R
\name{new.structure}
\alias{new.structure}
\title{Create new Structure object}
\usage{
new.structure(pdb.file)
}
\arguments{
\item{pdb.file}{The PDB structure file to use}
}
\value{
a PDB structure object
}
\description{
Constructor for a PDB structure object.
}
\details{
The resulting object offers the following methods:
\itemize{
\item get.sequence(chain) Returns the amino acid sequence of the given chain
\item get.aa(chain,pos) Returns a data.frame with the atomic information for 
   the amino acid determined by the given chain and position.
\item get.chains() Returns a vector detailing the chain indentifiers in the structure
\item get.atom(chain,pos,name) Returns the atomic information associated with the 
   given chain, amino acid position, and atom name.
\item subcomplex.combos(chain) !Deprecated! Does not work with NMR structures
\item get.chain.info() Returns a dataframe listing which chain corresponds to 
    which protein.
}
}
