% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackdrawer.R
\name{new.trackdrawer}
\alias{new.trackdrawer}
\title{New Trackdrawer object}
\usage{
new.trackdrawer(l, nox = FALSE)
}
\arguments{
\item{l}{the length of the tracks (i.e. the number of amino acids)}

\item{nox}{no x-axis. Removes the x-axis and decreases the bottom margin to 0. This
option is for use with the \code{layout()} function, to support adding tracks on top
of an existing graph that uses amino acid postion as the x-axis.}
}
\value{
an object of type TrackDrawer.
}
\description{
This constructor creates a new Trackdrawer object. The object can be used to draw
a plot containing amino-acid resolution structural information tracks, such as 
conservation, secondary structure, solvent accessibility or burial.
}
\details{
The object has the following methods:
\itemize{
  \item \code{add.track(values,label,col,minVal=0,maxVal=max(values))} adds a new numerical
    track, which will be visualized in heatmap style. \code{values} are the numerical
    values for the heatmap. \code{label} is the label that will be shown on the y-axis
    for this track. \code{col} is the color that will be used to generate the color
    ramp (from white to \code{col}). \code{minVal} is optional and is the minimum value that
    will be mapped to the color white (default 0). \code{maxVal} is also optional and
    is the maximum value in the scale, which will be mapped to \code{col}.
  \item \code{add.constrack(values)} adds a new conservation track, which will be 
    visualized as a barplot. \code{values} are the conservation values to use.
  \item \code{add.ss.track(values)} adds a new secondary structure track, which 
    will be visualized using spriral and arrow symbols. \code{values} are the 
    character values representing the secondary structure type for each amino acid.
    Permissible values are: "AlphaHelix","H","G","I" for alpha helices, "Strand","E"
    for beta strands and "Disorder" for disordered regions. Any other string is 
    interpreted as no specific secondary structure.
  \item \code{draw()} draws the plot
  \item \code{num.tracks()} returns the number of tracks currently added to the plot.
}
}
\examples{
\dontrun{
td <- new.trackdrawer(50)
td$add.track(runif(50,0,1),"SASA","blue")
td$add.constrack(runif(50,0,11))
td$add.ss.track(sample(c("AlphaHelix","Strand",""),50,replace=TRUE))
td$draw()
}
}
