% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologs.R
\name{pdb.informative}
\alias{pdb.informative}
\title{Select smallest informative subset of PDB structures}
\usage{
pdb.informative(pdb.table)
}
\arguments{
\item{pdb.table}{The result of \code{find.pdbs()}}
}
\value{
a vector with the IDs of the selected PDB structures.
}
\description{
Given the results of \code{find.pdbs()}, this function finds the smallest informative
subset among them. That is, the smallest set of PDB structures, that still represent all
available interaction partners.
}
