% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStrucFeats.R
\name{query.alphafold}
\alias{query.alphafold}
\title{Query Alphafold}
\usage{
query.alphafold(uniprot.acc, pdb.file = paste0(uniprot.acc, "_af.pdb"))
}
\arguments{
\item{uniprot.acc}{the Uniprot accession to query}

\item{pdb.file}{the name of the file to which the output will be written.}
}
\value{
the name of the output file.
}
\description{
Queries the EMBL Alphafold database for predicted protein structure data for
a given database accession and writes the result to file
}
\examples{
\dontrun{
#with a pre-determined output file
outfile <- "sumo_conjugase_structure.pdb"
query.pdb("3UIP",outfile)

#with an autogenerated output file
outfile <- query.pdb("3UIP")
}
}
