% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{retrieve.async.result.file}
\alias{retrieve.async.result.file}
\title{Retrieve result file from asynchronous job}
\usage{
retrieve.async.result.file(jobID, ext = "png")
}
\arguments{
\item{jobID}{the ID of the job.}

\item{ext}{the file extension. Defaults to "png".}
}
\value{
\code{TRUE} if successful, otherwise \code{FALSE}.
}
\description{
Retrieves the output file from an asynchronous job if it exists.
The file is copied to the working directory, so that OpenCPU can expose it
on the webservice.
}
