% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStrucFeats.R
\name{run.dssp}
\alias{run.dssp}
\title{Run DSSP on a PDB structure}
\usage{
run.dssp(pdb.file)
}
\arguments{
\item{pdb.file}{The PDB file on which to run DSSP}
}
\value{
a \code{data.frame} with the output of DSSP
}
\description{
Calculates secondary structure information from a PDB file using the
external software DSSP, which must be installed and available via $PATH
}
\examples{
\dontrun{
dssp.out <- run.dssp(pdb.file)
}
}
