% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStrucFeats.R
\name{run.sasa}
\alias{run.sasa}
\title{Run FreeSASA on a PDB structure}
\usage{
run.sasa(pdb.file)
}
\arguments{
\item{pdb.file}{The PDB file on which to run FreeSASA}
}
\value{
a \code{data.frame} with the output of FreeSASA
}
\description{
Calculates solvent accessible surface area on a given structure using the 
external software FreeSASA, which must be installed and available via $PATH.
}
\examples{
\dontrun{
sasa <- run.sasa("3UIP.pdb")
}
}
