% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStrucFeats.R
\name{subcomplex.combos}
\alias{subcomplex.combos}
\title{Generate PDB files for complex components}
\usage{
subcomplex.combos(pdb.file, chain.sets)
}
\arguments{
\item{pdb.file}{the input PDB file for the complex}

\item{chain.sets}{a list of character vectors, detailing the combinations of chains to 
produce. E.g. \code{list("A",c("A","B"),c("A","C"))} produces one file for 
chain A, one for the combination of chains A and B, and one for the combination
of chains A and C.}
}
\value{
a vector of file names corresponding to the elements of \code{chain.sets}
}
\description{
This function finds the component chains of a given PDB structure and generates
new PDB files for any desired set of combinations of these components
}
\examples{
\dontrun{
subcplx.files <- sub.complex.combos("3UIP.pdb",chain.sets=list("A",c("A","B"),c("A","C")))
}
}
