% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genophenogram.R
\name{genophenogram}
\alias{genophenogram}
\title{Draw genophenogram plot}
\usage{
genophenogram(wt.aa, pos, mut.aa, score, syn.med, stop.med, error = NULL,
  a = 0, grayBack = FALSE, img.width = 12, tracks = NULL)
}
\arguments{
\item{wt.aa}{wildtype amino acid sequence as a vector of single characters.}

\item{pos}{vector of amino acid positions}

\item{mut.aa}{vector of mutant AAs}

\item{score}{vector of scores}

\item{error}{vector of stderr values}

\item{a}{bezier transformation intensity (with -0.5 <= a <= 0.5)}

\item{grayBack}{draw a gray background for incomplete maps}

\item{img.width}{optional parameter to inform the drawing function of the chosen image width,
allowing it to adjust the size of the legend}

\item{tracks}{an optional trackdrawer object to add structural information to the plot}
}
\description{
Draws a genophenogram plot from given data
}
